+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 - 1.1.0
     - Added Reversible transformer interface
-    - Max Abs and Z Score scaler are now Reversible
+    - MaxAbs, Z Score, and Robust scalers are now Reversible
 
 - 1.0.0
     - No changes